+++ playgrounds/vite/src/app.tsx
@@ -1,6 +1,6 @@
 export function App() {
   return (
-    <div className="m-3 p-3 border">
+    <div className="m-3 p-3 border max-w-screen-sm">
       <h1 className="text-blue-500">Hello World</h1>
       <div className="-inset-x-full -inset-y-full -space-x-full -space-y-full -inset-full"></div>
     </div>
